backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field
phonenumbers
django-cors-headers
drf-yasg
coverage